More work on requested changes
